Move MySiteCardAndItem item related classes to items package
