+++ WordPress/src/main/java/org/wordpress/android/ui/mysite/MySiteViewModel.kt
@@ -45,24 +45,24 @@ import org.wordpress.android.fluxc.store.QuickStartStore.QuickStartTaskType
 import org.wordpress.android.modules.BG_THREAD
 import org.wordpress.android.modules.UI_THREAD
 import org.wordpress.android.ui.PagePostCreationSourcesDetail.STORY_FROM_MY_SITE
-import org.wordpress.android.ui.mysite.ListItemAction.ACTIVITY_LOG
-import org.wordpress.android.ui.mysite.ListItemAction.ADMIN
-import org.wordpress.android.ui.mysite.ListItemAction.BACKUP
-import org.wordpress.android.ui.mysite.ListItemAction.COMMENTS
-import org.wordpress.android.ui.mysite.ListItemAction.DOMAINS
-import org.wordpress.android.ui.mysite.ListItemAction.JETPACK_SETTINGS
-import org.wordpress.android.ui.mysite.ListItemAction.MEDIA
-import org.wordpress.android.ui.mysite.ListItemAction.PAGES
-import org.wordpress.android.ui.mysite.ListItemAction.PEOPLE
-import org.wordpress.android.ui.mysite.ListItemAction.PLAN
-import org.wordpress.android.ui.mysite.ListItemAction.PLUGINS
-import org.wordpress.android.ui.mysite.ListItemAction.POSTS
-import org.wordpress.android.ui.mysite.ListItemAction.SCAN
-import org.wordpress.android.ui.mysite.ListItemAction.SHARING
-import org.wordpress.android.ui.mysite.ListItemAction.SITE_SETTINGS
-import org.wordpress.android.ui.mysite.ListItemAction.STATS
-import org.wordpress.android.ui.mysite.ListItemAction.THEMES
-import org.wordpress.android.ui.mysite.ListItemAction.VIEW_SITE
+import org.wordpress.android.ui.mysite.items.listitem.ListItemAction.ACTIVITY_LOG
+import org.wordpress.android.ui.mysite.items.listitem.ListItemAction.ADMIN
+import org.wordpress.android.ui.mysite.items.listitem.ListItemAction.BACKUP
+import org.wordpress.android.ui.mysite.items.listitem.ListItemAction.COMMENTS
+import org.wordpress.android.ui.mysite.items.listitem.ListItemAction.DOMAINS
+import org.wordpress.android.ui.mysite.items.listitem.ListItemAction.JETPACK_SETTINGS
+import org.wordpress.android.ui.mysite.items.listitem.ListItemAction.MEDIA
+import org.wordpress.android.ui.mysite.items.listitem.ListItemAction.PAGES
+import org.wordpress.android.ui.mysite.items.listitem.ListItemAction.PEOPLE
+import org.wordpress.android.ui.mysite.items.listitem.ListItemAction.PLAN
+import org.wordpress.android.ui.mysite.items.listitem.ListItemAction.PLUGINS
+import org.wordpress.android.ui.mysite.items.listitem.ListItemAction.POSTS
+import org.wordpress.android.ui.mysite.items.listitem.ListItemAction.SCAN
+import org.wordpress.android.ui.mysite.items.listitem.ListItemAction.SHARING
+import org.wordpress.android.ui.mysite.items.listitem.ListItemAction.SITE_SETTINGS
+import org.wordpress.android.ui.mysite.items.listitem.ListItemAction.STATS
+import org.wordpress.android.ui.mysite.items.listitem.ListItemAction.THEMES
+import org.wordpress.android.ui.mysite.items.listitem.ListItemAction.VIEW_SITE
 import org.wordpress.android.ui.mysite.MySiteCardAndItem.Card.DomainRegistrationCard
 import org.wordpress.android.ui.mysite.MySiteCardAndItem.DynamicCard
 import org.wordpress.android.ui.mysite.SiteDialogModel.AddSiteIconDialogModel
@@ -103,6 +103,8 @@ import org.wordpress.android.ui.mysite.dynamiccards.DynamicCardMenuViewModel.Dyn
 import org.wordpress.android.ui.mysite.dynamiccards.DynamicCardMenuViewModel.DynamicCardMenuInteraction.Pin
 import org.wordpress.android.ui.mysite.dynamiccards.DynamicCardMenuViewModel.DynamicCardMenuInteraction.Unpin
 import org.wordpress.android.ui.mysite.dynamiccards.DynamicCardsSource
+import org.wordpress.android.ui.mysite.items.listitem.ListItemAction
+import org.wordpress.android.ui.mysite.items.SiteItemsBuilder
 import org.wordpress.android.ui.mysite.quickactions.QuickActionsCardBuilder
 import org.wordpress.android.ui.mysite.quickstart.QuickStartCardBuilder
 import org.wordpress.android.ui.pages.SnackbarMessageHolder
