+++ WordPress/src/main/java/org/wordpress/android/ui/mysite/items/categoryheader/MySiteCategoryItemViewHolder.kt
@@ -1,8 +1,9 @@
-package org.wordpress.android.ui.mysite
+package org.wordpress.android.ui.mysite.items.categoryheader
 
 import android.view.ViewGroup
 import org.wordpress.android.databinding.MySiteCategoryHeaderItemBinding
 import org.wordpress.android.ui.mysite.MySiteCardAndItem.Item.CategoryHeaderItem
+import org.wordpress.android.ui.mysite.MySiteItemViewHolder
 import org.wordpress.android.ui.utils.UiHelpers
 import org.wordpress.android.util.viewBinding
