+++ WordPress/src/main/java/org/wordpress/android/ui/mysite/items/listitem/MySiteListItemViewHolder.kt
@@ -1,7 +1,9 @@
-package org.wordpress.android.ui.mysite
+package org.wordpress.android.ui.mysite.items.listitem
 
 import android.view.ViewGroup
 import org.wordpress.android.databinding.MySiteItemBlockBinding
+import org.wordpress.android.ui.mysite.MySiteCardAndItem.Item.ListItem
+import org.wordpress.android.ui.mysite.MySiteItemViewHolder
 import org.wordpress.android.ui.utils.UiHelpers
 import org.wordpress.android.util.viewBinding
 
@@ -9,7 +11,7 @@ class MySiteListItemViewHolder(
     parent: ViewGroup,
     private val uiHelpers: UiHelpers
 ) : MySiteItemViewHolder<MySiteItemBlockBinding>(parent.viewBinding(MySiteItemBlockBinding::inflate)) {
-    fun bind(cardAndItem: MySiteCardAndItem.Item.ListItem) = with(binding) {
+    fun bind(cardAndItem: ListItem) = with(binding) {
         uiHelpers.setImageOrHide(mySiteItemPrimaryIcon, cardAndItem.primaryIcon)
         uiHelpers.setImageOrHide(mySiteItemSecondaryIcon, cardAndItem.secondaryIcon)
         uiHelpers.setTextOrHide(mySiteItemPrimaryText, cardAndItem.primaryText)
